Fixed compiler warning C4838 'conversion from size_t to LONG requires a narrowing conversion'.
BUG=angleproject:1119

Reviewed-on: https://chromium-review.googlesource.com/292156
Tested-by: Cooper Partin <coopp@microsoft.com>

+++ util/win32/Win32Window.cpp
@@ -443,7 +443,7 @@ bool Win32Window::initialize(const std::string &name, size_t width, size_t heigh
     DWORD parentStyle = WS_CAPTION | WS_THICKFRAME | WS_MINIMIZEBOX | WS_MAXIMIZEBOX | WS_SYSMENU;
     DWORD parentExtendedStyle = WS_EX_APPWINDOW;
 
-    RECT sizeRect = { 0, 0, width, height };
+    RECT sizeRect = { 0, 0, static_cast<LONG>(width), static_cast<LONG>(height) };
     AdjustWindowRectEx(&sizeRect, parentStyle, FALSE, parentExtendedStyle);
 
     mParentWindow = CreateWindowExA(parentExtendedStyle, mParentClassName.c_str(), name.c_str(), parentStyle, CW_USEDEFAULT, CW_USEDEFAULT,
